Create new module for clear all completed tasks functionality
